LayoutTests/imported/w3c/web-platform-tests/preload/preload-referrer-policy-subresource-header.tentative_isCrossOriginPreload=true&isCrossOriginResource=false-expected.txt

The referrerpolicy attribute on Link header should be ignored for subresources to prevent cross-origin referrer leakage


FAIL referrer policy ( -> , cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy ( -> no-referrer, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy ( -> same-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy ( -> origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy ( -> origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy ( -> strict-origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy ( -> unsafe-url, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
PASS referrer policy (no-referrer -> , cross-origin, same-origin)
PASS referrer policy (no-referrer -> no-referrer, cross-origin, same-origin)
PASS referrer policy (no-referrer -> same-origin, cross-origin, same-origin)
PASS referrer policy (no-referrer -> origin, cross-origin, same-origin)
PASS referrer policy (no-referrer -> origin-when-cross-origin, cross-origin, same-origin)
PASS referrer policy (no-referrer -> strict-origin-when-cross-origin, cross-origin, same-origin)
PASS referrer policy (no-referrer -> unsafe-url, cross-origin, same-origin)
PASS referrer policy (same-origin -> , cross-origin, same-origin)
PASS referrer policy (same-origin -> no-referrer, cross-origin, same-origin)
PASS referrer policy (same-origin -> same-origin, cross-origin, same-origin)
PASS referrer policy (same-origin -> origin, cross-origin, same-origin)
PASS referrer policy (same-origin -> origin-when-cross-origin, cross-origin, same-origin)
PASS referrer policy (same-origin -> strict-origin-when-cross-origin, cross-origin, same-origin)
PASS referrer policy (same-origin -> unsafe-url, cross-origin, same-origin)
FAIL referrer policy (origin -> , cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> no-referrer, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> same-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> strict-origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> unsafe-url, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin-when-cross-origin -> , cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin-when-cross-origin -> no-referrer, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin-when-cross-origin -> same-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin-when-cross-origin -> origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin-when-cross-origin -> origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin-when-cross-origin -> strict-origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin-when-cross-origin -> unsafe-url, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (strict-origin-when-cross-origin -> , cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (strict-origin-when-cross-origin -> no-referrer, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (strict-origin-when-cross-origin -> same-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (strict-origin-when-cross-origin -> origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (strict-origin-when-cross-origin -> origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (strict-origin-when-cross-origin -> strict-origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (strict-origin-when-cross-origin -> unsafe-url, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (unsafe-url -> , cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=true&isCrossOriginResource=false"
FAIL referrer policy (unsafe-url -> no-referrer, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=true&isCrossOriginResource=false"
FAIL referrer policy (unsafe-url -> same-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=true&isCrossOriginResource=false"
FAIL referrer policy (unsafe-url -> origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=true&isCrossOriginResource=false"
FAIL referrer policy (unsafe-url -> origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=true&isCrossOriginResource=false"
FAIL referrer policy (unsafe-url -> strict-origin-when-cross-origin, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=true&isCrossOriginResource=false"
FAIL referrer policy (unsafe-url -> unsafe-url, cross-origin, same-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=true&isCrossOriginResource=false"
